Arps decline prediction

Create a data frame of decline predictions according to the class of 'decline_lst' and 'time_lst' objects

decline_predict(decline_lst, time_lst)

Arguments

decline_lst
a list object of class 'decline'

time_lst
a list object of class 'time'

Value

a data frame of decline estimates according to the class of 'decline_lst' and 'time_lst' objects

References

Arps JJ (1945). “Analysis of Decline Curves.” Transactions of the AIME, 160(01), 228--247. ISSN 0081-1696, doi: 10.2118/945228-G , https://doi.org/10.2118/945228-G.

Robertson S (1988). “Generalized Hyperbolic Equation.”

Examples

decline_param_1 <- decline_param(input_unit = "Field", output_unit = "Field",
fluid = "oil",
model = "exponential", qi = 1000, Di = 0.15, b = 0, q_abnd = NULL)
decline_time_1 <- decline_time(c(1:7300), unit = "day")
decline_predict_1 <- decline_predict(decline_param_1, decline_time_1)

head(decline_predict_1)

#>         Date Time_(day) q_(bbl/day)   Q_(bbl) D_(1/day) Beta
#> 1 2020-05-13          1    860.7080  928.6135      0.15    0
#> 2 2020-05-14          2    740.8182 1727.8785      0.15    0
#> 3 2020-05-15          3    637.6282 2415.8123      0.15    0
#> 4 2020-05-16          4    548.8116 3007.9224      0.15    0
#> 5 2020-05-17          5    472.3666 3517.5563      0.15    0
#> 6 2020-05-18          6    406.5697 3956.2023      0.15    0

decline_param_2 <- decline_param(input_unit = "Field", output_unit = "SI",
fluid = "oil",
model = "hyperbolic", qi = 15000, Di = 0.1, b = 0.95, q_abnd = 200)
decline_time_2 <- decline_time(seq(as.Date("2016/04/15"), by = "days",
length.out = 3600), unit = "date")
decline_predict_2 <- decline_predict(decline_param_2, decline_time_2)

head(decline_predict_2)

#>         Date Time_(day) q_(m3/day)    Q_(m3)  D_(1/day) Beta time_abnd_(days)
#> 1 2016-04-15          1   2167.529  2272.797 0.09132420 0.95         625.6604
#> 2 2016-04-16          2   1985.776  4346.863 0.08403361 0.95         625.6604
#> 3 2016-04-17          3   1831.549  6253.500 0.07782101 0.95         625.6604
#> 4 2016-04-18          4   1699.074  8017.196 0.07246377 0.95         625.6604
#> 5 2016-04-19          5   1584.082  9657.465 0.06779661 0.95         625.6604
#> 6 2016-04-20          6   1483.349 11190.105 0.06369427 0.95         625.6604
#>   EUR_(m3)
#> 1 92608.54
#> 2 92608.54
#> 3 92608.54
#> 4 92608.54
#> 5 92608.54
#> 6 92608.54